Reorganized Layout
 - Created separate Lecture and Lab folder.
- Threw in existing daily lecture and lab folders in their respective folders.
- Renamed title of daily lecture and lab folders to simplify searches for specific code.

The change in layout should make it easier to access both labs and lectures, as each is now kept in separate folders. Also, the change in daily folder names should improve accessibility when searching for specific code, as they are titled after their respective names on eCampus' Course Materials page.